Add CSS class to enable fallback links
Adding the class `jslghtbx-fallback-link` to a surrounding link element
will now still trigger the Lightbox on click, but also allow a fallback
mechanism for non-JavaScript-enabled browsing to still present those
users with a link to e.g. the full-size image. It also enables the
expected actions for both JS and non-JS browsing when e.g.
middle-clicking to open the full-size link in the background and such.